- CLean up and full implementation

+++ src/audio_processing.h
@@ -48,7 +48,7 @@ void processAudioData(int16_t *data, uint16_t num_samples);
 *	put the invoking thread into sleep until it can process the audio datas
 */
 void wait_for_data(void);
-int16_t get_command(void);
+uint8_t get_command(void);
 void compute_command(void);
 
 void listen_init(void);
